refactor(cli): remove skeleton alias, rename default pattern to copy
Per Carl's #319 review:
- Drop the deprecated `skeleton` command entirely (SkeletonCommand, its CLI
  registration, and the alias test) instead of shipping a deprecated alias.
  It never released, so there is nothing to deprecate.
- Rename the bundled pattern from default/DefaultPattern to copy/CopyPattern
  (module="copy", flepimop2.pattern.copy), naming the module by what it does.

Updates the CLI default pattern reference, the ABC and command docstrings, the
CHANGELOG, and the tests (targetless/configless SkeletonCommand cases move to
PatternCommand). templates/skeleton stays as the template-tree source.
