+++ src/flepimop2/cli/_cli.py
@@ -29,7 +29,6 @@ from flepimop2.cli._pattern_command import PatternCommand
 from flepimop2.cli._process_command import ProcessCommand
 from flepimop2.cli._register_command import register_command
 from flepimop2.cli._simulate_command import SimulateCommand
-from flepimop2.cli._skeleton_command import SkeletonCommand
 
 _FLEPIMOP2_VERSION: Final[str] = version("flepimop2")
 
@@ -50,7 +49,6 @@ register_command(PatchCommand, cli)
 register_command(PatternCommand, cli)
 register_command(ProcessCommand, cli)
 register_command(SimulateCommand, cli)
-register_command(SkeletonCommand, cli)
 
 # Register subgroups
 cli.add_command(job_group)
